requirements.txt
beautifulsoup4==4.12.3
numpy==2.1.0
Pillow==10.4.0
PyExecJS==1.5.1
Requests==2.32.3
